line bot + line pay + ec pay 整條 95% finished 基本上已經可 demo

+++ src/main/java/com/eventra/exhibition/model/ExhibitionDTO.java
@@ -23,7 +23,7 @@ public class ExhibitionDTO {
 	private String location;
 	private String description;
 	private ExhibitorDTO exhibitor; // 帶入 id 後拼接展商主頁網址
-	private Long totalCommentCount;
+	private Integer totalCommentCount;
 	
 	public String getPhotoLandscape() {
 		return photoLandscape;
@@ -103,10 +103,10 @@ public class ExhibitionDTO {
 	public void setCheapestTicketPrice(Integer cheapestTicketPrice) {
 		this.cheapestTicketPrice = cheapestTicketPrice;
 	}
-	public Long getTotalCommentCount() {
+	public Integer getTotalCommentCount() {
 		return totalCommentCount;
 	}
-	public void setTotalCommentCount(Long totalCommentCount) {
+	public void setTotalCommentCount(Integer totalCommentCount) {
 		this.totalCommentCount = totalCommentCount;
 	}
 }
